LPS-193868 Move package, package should follow Bundle-SymbolicName
