+++ modules/dxp/apps/akismet/akismet-web/src/main/java/com/liferay/akismet/web/internal/constants/ModerationPortletKeys.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.akismet.internal.constants;
+package com.liferay.akismet.web.internal.constants;
 
 /**
  * @author Jamie
