+++ modules/dxp/apps/akismet/akismet-web/src/main/java/com/liferay/akismet/web/internal/portlet/ModerationPortlet.java
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.akismet.internal.portlet;
+package com.liferay.akismet.web.internal.portlet;
 
 import com.liferay.akismet.client.AkismetClient;
-import com.liferay.akismet.internal.constants.ModerationPortletKeys;
+import com.liferay.akismet.web.internal.constants.ModerationPortletKeys;
 import com.liferay.message.boards.exception.NoSuchMessageException;
 import com.liferay.message.boards.exception.RequiredMessageException;
 import com.liferay.message.boards.model.MBMessage;
