+++ modules/dxp/apps/akismet/akismet-web/src/main/java/com/liferay/akismet/web/internal/portlet/action/AkismetEditMessageMVCActionCommand.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.akismet.internal.portlet.action;
+package com.liferay.akismet.web.internal.portlet.action;
 
 import com.liferay.akismet.client.AkismetClient;
 import com.liferay.message.boards.constants.MBPortletKeys;
