+++ modules/dxp/apps/akismet/akismet-web/src/main/resources/META-INF/resources/dynamic_include/init.jsp
@@ -13,7 +13,7 @@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
-<%@ page import="com.liferay.akismet.internal.security.permission.MBResourcePermission" %><%@
+<%@ page import="com.liferay.akismet.web.internal.security.permission.MBResourcePermission" %><%@
 page import="com.liferay.message.boards.constants.MBCategoryConstants" %><%@
 page import="com.liferay.message.boards.model.MBCategory" %><%@
 page import="com.liferay.message.boards.model.MBMessage" %><%@
