+++ modules/dxp/apps/akismet/akismet-web/src/main/resources/META-INF/resources/init.jsp
@@ -15,7 +15,7 @@ taglib uri="http://liferay.com/tld/portlet" prefix="liferay-portlet" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
-<%@ page import="com.liferay.akismet.internal.util.ModerationUtil" %><%@
+<%@ page import="com.liferay.akismet.web.internal.util.ModerationUtil" %><%@
 page import="com.liferay.message.boards.exception.NoSuchMessageException" %><%@
 page import="com.liferay.message.boards.exception.RequiredMessageException" %><%@
 page import="com.liferay.message.boards.model.MBMessage" %><%@
